Fix BookRepositoryImpl: Long type mismatches
1. `totalChars = 0` was Int but BookEntity.totalChars is Long -> changed to `0L`
2. `totalChars += charCount` was Long += Int which doesn't compile in Kotlin
   -> changed to `totalChars += charCount.toLong()`

+++ core/data/src/main/java/com/epubaudioreader/core/data/repository/BookRepositoryImpl.kt
@@ -116,7 +116,7 @@ class BookRepositoryImpl @Inject constructor(
                     description = parsedEpub.metadata.description,
                     filePath = filePath,
                     totalChapters = parsedEpub.spine.size,
-                    totalChars = 0,
+                    totalChars = 0L,
                     fileSize = fileSize,
                     hash = hash
                 )
@@ -151,7 +151,7 @@ class BookRepositoryImpl @Inject constructor(
 
                 for ((index, chapterData) in chapterDataList.withIndex()) {
                     val charCount = chapterData.text.length
-                    totalChars += charCount
+                    totalChars += charCount.toLong()
 
                     // Save chapter text to filesystem (I/O outside transaction)
                     val path = storageManager.saveChapterText(
